% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLUS_clusteringVaryLj.R
\name{clusteringVaryLj}
\alias{clusteringVaryLj}
\title{clusteringVaryLj}
\usage{
clusteringVaryLj(object, o = 1, nkeep = 1, nCent = 3)
}
\arguments{
\item{object}{A \code{\link{varyLjBFA}} or \code{\link{VAR1varyLjBFA}} model object on which sptial temporal trends clustering is based.}

\item{o}{A positive integer less than or equal to O (the number of observation types) indicating the single observation type that we want to perform clustering for.}

\item{nkeep}{A positive integer less than or equal to NKeep (the number of post-burn-in post-thinned MCMC iterations) indicating the number of selected kept 
post-burn-in MCMC iterations whose posterior parameter estimates our spatial temporal trends clustering will be based on.}

\item{nCent}{A positive integer indicating the number of clusters for k-means clustering.}
}
\value{
\code{clusteringVaryLj} returns an object from R's \code{kmeans()} function, which is a list with elements \code{cluster}, \code{centers},
\code{totss}, \code{withinss}, \code{tot.withinss}, \code{betweenss}, \code{size}, \code{iter}, and \code{ifault}.
}
\description{
Cluster spatial points into regions with similar temporal trajectories from a \code{\link{varyLjBFA}} model.
}
\references{
Yifan Cheng
}
