% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Utilities.R
\name{is.FixedLbfa}
\alias{is.FixedLbfa}
\title{is.FixedLbfa}
\usage{
is.FixedLbfa(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
\code{is.FixedLbfa} returns a logical, depending on whether the object is of class \code{\link{FixedLbfa}}.
}
\description{
\code{is.FixedLbfa} is a general test of an object being interpretable as a
\code{\link{FixedLbfa}} object.
}
\details{
The \code{\link{FixedLbfa}} class is defined as the regression object that
 results from the \code{\link{bfaFixedL}} regression function.
}
\examples{
###Load pre-computed results
data(reg.FixedLbfa)

###Test function
is.FixedLbfa(reg.FixedLbfa)

}
