% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Utilities.R
\name{is.FixedLbfaVAR1}
\alias{is.FixedLbfaVAR1}
\title{is.FixedLbfaVAR1}
\usage{
is.FixedLbfaVAR1(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
\code{is.FixedLbfaVAR1} returns a logical, depending on whether the object is of class \code{\link{FixedLbfaVAR1}}.
}
\description{
\code{is.FixedLbfaVAR1} is a general test of an object being interpretable as a
\code{\link{FixedLbfaVAR1}} object.
}
\details{
The \code{\link{FixedLbfaVAR1}} class is defined as the regression object that
 results from the \code{\link{VAR1bfaFixedL}} regression function.
}
\examples{
###Load pre-computed results
data(reg.FixedLbfaVAR1)

###Test function
is.FixedLbfaVAR1(reg.FixedLbfaVAR1)

}
