% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Utilities.R
\name{is.VAR1varyLjBFA}
\alias{is.VAR1varyLjBFA}
\title{is.VAR1varyLjBFA}
\usage{
is.VAR1varyLjBFA(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
\code{is.VAR1varyLjBFA} returns a logical, depending on whether the object is of class \code{\link{VAR1varyLjBFA}}.
}
\description{
\code{is.VAR1varyLjBFA} is a general test of an object being interpretable as a
\code{\link{VAR1varyLjBFA}} object.
}
\details{
The \code{\link{VAR1varyLjBFA}} class is defined as the regression object that
 results from the \code{\link{VAR1bfaVaryingLjs}} regression function.
}
\examples{
###Load pre-computed results
data(reg.VAR1varyLjBFA)

###Test function
is.varyLjBFA(reg.VAR1varyLjBFA)

}
