% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Utilities.R
\name{is.varyLjBFA}
\alias{is.varyLjBFA}
\title{is.varyLjBFA}
\usage{
is.varyLjBFA(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
\code{is.varyLjBFA} returns a logical, depending on whether the object is of class \code{\link{varyLjBFA}}.
}
\description{
\code{is.varyLjBFA} is a general test of an object being interpretable as a
\code{\link{varyLjBFA}} object.
}
\details{
The \code{\link{varyLjBFA}} class is defined as the regression object that
 results from the \code{\link{bfaVaryingLjs}} regression function.
}
\examples{
###Load pre-computed results
data(reg.varyLjBFA)

###Test function
is.varyLjBFA(reg.varyLjBFA)

}
