% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRED_predictNewTime.R
\name{predictNewTime}
\alias{predictNewTime}
\title{predictNewTime}
\usage{
predictNewTime(
  object,
  NewTimes,
  NewX = NULL,
  NewTrials = NULL,
  Verbose = TRUE,
  seed = 54,
  ...
)
}
\arguments{
\item{object}{A \code{\link{FixedLbfa}} or \code{\link{varyLjBFA}} model object for which predictions
are desired from.}

\item{NewTimes}{A numeric vector including desired time point(s) for prediction. If \code{include.time = equalTimeDist = TRUE} when getting this function's input \code{object} from the main function \code{bfa_sp} 
and the pre-normalized time distance does not equal 1, then we should standardize the new time points accordingly. \code{NewTimes} should be the standardized desired time point(s) for prediction.}

\item{NewX}{A matrix including covariates at times \code{NewTimes} for prediction. 
\code{NewX} must have dimension \code{(M x O x NNewVisits) x P}. Where \code{NNewVisits} is the number of temporal 
locations being predicted. The default sets \code{NewX} to \code{NULL}, which assumes that the covariates for all predictions 
are the same as the final time point.}

\item{NewTrials}{An array indicating the trials for categorical predictions. The array must have dimension \code{M x C x NNewVisits}
and contain only non-negative integers. The default sets \code{NewTrials} to \code{NULL}, which assumes that trials for all predictions
are the same as the final time point.}

\item{Verbose}{A boolean logical indicating whether progress should be output.}

\item{seed}{An integer value used to set the seed for the random number generator
(default = 54).}

\item{...}{other arguments.}
}
\value{
\code{predictNewTime} returns a list containing the following objects.

  \describe{

  \item{\code{Eta}}{A \code{list} containing \code{NNewVisits} matrices, one for each new time prediction. Each matrix is dimension \code{NKeep x K}, where
  \code{K} is the number of latent factors. Each matrix contains posterior samples obtained by Bayesian krigging.}

  \item{\code{Y}}{A \code{list} containing \code{NNewVisits} posterior predictive distribution matrices. 
  Each matrix is of dimension \code{NKeep x (M * O)}, where \code{M} is the number of spatial locations and \code{O} the number of observation types.
  Each matrix is obtained through Bayesian krigging.}

  }
}
\description{
Predicts future observation(s) from a \code{\link{FixedLbfa}} or \code{\link{varyLjBFA}} model.
}
\details{
\code{predictNewTime} uses Bayesian krigging to predict vectors at future
 time points. The function returns the krigged factors (\code{Eta}) and also the observed outcomes (\code{Y}).
}
\references{
Yifan Cheng

Berchuck, S. I., Janko, M., Medeiros, F. A., Pan, W., & Mukherjee, S. (2021). Bayesian Non-Parametric Factor Analysis for Longitudinal Spatial Surfaces. Bayesian Analysis, 17(2), 1–30.
}
